backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
cohere>=5.0.0
pinecone-client==3.0.0
pillow>=10.0.0
python-dotenv>=1.0.0

